split riscv modules in crates so other crates can have minimal deps (#2747)
