+++ openvm/Cargo.toml
@@ -46,7 +46,7 @@ openvm-stark-backend = { git = "https://github.com/powdr-labs/stark-backend.git"
 
 powdr-ast.workspace = true
 powdr-number.workspace = true
-powdr-riscv.workspace = true
+powdr-riscv-elf.workspace = true
 powdr-autoprecompiles.workspace = true
 powdr-constraint-solver.workspace = true
