+++ riscv-elf/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "powdr-riscv-elf"
+description = "powdr RISCV ELF utils"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+powdr-isa-utils.workspace = true
+powdr-riscv-types.workspace = true
+powdr-syscalls.workspace = true
+
+gimli = "0.31"
+goblin = "0.8"
+itertools = "0.13"
+log = "0.4.27"
+raki = "0.1.4"
+thiserror = "1.0"
+
+[lints]
+workspace = true
+
+[lib]
+bench = false # See https://github.com/bheisler/criterion.rs/issues/458
