+++ riscv-types/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "powdr-riscv-types"
+description = "powdr RISCV types and traits"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+powdr-isa-utils.workspace = true
+
+[lints]
+workspace = true
+
+[lib]
+bench = false # See https://github.com/bheisler/criterion.rs/issues/458
